feat: Update `ResizablePanel`
