+++ src/stories/atom/ResizablePanel/index.tsx
@@ -1,27 +1,23 @@
-import { useState } from 'react';
 import { Resizable } from 're-resizable';
+import { useResizable } from '~/stories/atom/ResizablePanel/useResizable';
 
 interface ResizablePanelProps {
   className?: string;
 }
 
 export const ResizablePanel = (props: ResizablePanelProps) => {
-  const [height, setHeight] = useState<number>(200);
-
-  const handleResizeStop = (
-    e: MouseEvent | TouchEvent,
-    direction: string,
-    ref: HTMLElement,
-    delta: { width: number; height: number }
-  ): void => {
-    setHeight(height + delta.height);
-  };
+  const { width, height, minHeight, maxHeight, handleResizeStop } = useResizable({
+    initialWidth: 300,
+    initialHeight: 200,
+    minHeight: 100,
+    maxHeight: 400,
+  });
 
   return (
     <Resizable
-      size={{ width: 300, height }}
-      minHeight={100}
-      maxHeight={400}
+      size={{ width, height }}
+      minHeight={minHeight}
+      maxHeight={maxHeight}
       enable={{
         top: true,
         right: false,
@@ -35,10 +31,7 @@ export const ResizablePanel = (props: ResizablePanelProps) => {
       onResizeStop={handleResizeStop}
       className="border !border-primary"
     >
-      <div className=" panel-content">
-        <p>This is a resizable panel with dynamic size.</p>
-        <p>You can add more content here.</p>
-      </div>
+      Sample with resizable vertical height
     </Resizable>
   );
 };
